+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Add relation with a file to scenario `security.filePermission`.
+
+### Changed
+- Lint JSON files in parallel through new script `lint-json.sh`, usable both locally and in CI.
+
+### Fixed
+- Fix file count and duplicate id check in `update-backup.sh`, add check for files without corresponding element.
 
 ## 0.0.31 - 2026-09-23